api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.52.0
sqlglot>=26.0.0
psycopg[binary,pool]>=3.2.0
pydantic-settings>=2.0.0
cachetools>=5.3.0
